+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/shared/_internal/createHubdleSharedExtensions.kt
@@ -1,19 +1,19 @@
 package com.javiersc.hubdle.project.extensions.shared._internal
 
 import com.javiersc.hubdle.project.extensions._internal.HubdleState
-import com.javiersc.hubdle.project.extensions.shared.features.HubdleGradleFeatureExtension
+import com.javiersc.hubdle.project.extensions.gradle.HubdleGradleExtension
+import com.javiersc.hubdle.project.extensions.gradle.HubdleGradlePluginExtension
+import com.javiersc.hubdle.project.extensions.gradle.HubdleGradleVersionCatalogExtension
 import com.javiersc.hubdle.project.extensions.shared.features.HubdleIntellijFeatureExtension
 import com.javiersc.hubdle.project.extensions.shared.features.HubdleJavaApplicationFeatureExtension
 import com.javiersc.hubdle.project.extensions.shared.features.HubdleJavaVersionFeatureExtension
 import com.javiersc.hubdle.project.extensions.shared.features.HubdleJvmToolchainFeatureExtension
-import com.javiersc.hubdle.project.extensions.shared.features.gradle.HubdleGradlePluginFeatureExtension
-import com.javiersc.hubdle.project.extensions.shared.features.gradle.HubdleGradleVersionCatalogFeatureExtension
 import com.javiersc.hubdle.project.extensions.shared.features.intellij.HubdleIntellijPluginFeatureExtension
 
 internal fun HubdleState.createHubdleSharedExtensions() {
-    createExtension<HubdleGradleFeatureExtension> {
-        createExtension<HubdleGradlePluginFeatureExtension>()
-        createExtension<HubdleGradleVersionCatalogFeatureExtension>()
+    createExtension<HubdleGradleExtension> {
+        createExtension<HubdleGradlePluginExtension>()
+        createExtension<HubdleGradleVersionCatalogExtension>()
     }
     createExtension<HubdleIntellijFeatureExtension> {
         createExtension<HubdleIntellijPluginFeatureExtension>()