+++ hubdle-gradle-plugin/testFunctional/kotlin/com/javiersc/hubdle/project/kotlin/jvm/KotlinJvmTest.kt
@@ -5,7 +5,6 @@ import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.file.shouldBeADirectory
 import io.kotest.matchers.file.shouldBeAFile
 import io.kotest.matchers.file.shouldNotExist
-import io.kotest.matchers.string.shouldContain
 import java.io.File
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
@@ -136,17 +135,6 @@ internal class KotlinJvmTest : GradleTestKitTest() {
             versionDir.resolve("sandbox-project-3.6.7-SNAPSHOT.pom").shouldBeAFile()
             versionDir.resolve("sandbox-project-3.6.7-SNAPSHOT-javadoc.jar").shouldBeAFile()
             versionDir.resolve("sandbox-project-3.6.7-SNAPSHOT-sources.jar").shouldBeAFile()
-        }
-    }
-
-    @Test
-    fun `publish gradle-plugin-1`() {
-        gradleTestKitTest("$basePath/publishing/gradle-plugin-1", name = projectName) {
-            withEnvironment(gpgMap)
-            withArgumentsFromTXT()
-            build()
-                .output
-                .shouldContain("Publications: [maven, pluginMaven, sandboxPluginMarkerMaven]")
         }
     }
 }